Adding imp config exclusion in addition to the others
Previous commit added prod, dev, sbx, and local config exclusions to the repo, but left out imp. This commit adds imp.

All of these files are flagged on an okta cred that is meant to be public for people to perform API testing, so there is no reason to flag on them.